requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
tomli==2.0.1
pytest-xdist==3.3.1
openpyxl==3.1.2